refactor: restruct the folder structure
