+++ src/components/button/copy-button/index.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/lib/components/ui/button";
+import { Button } from "@/components/ui/button";
 import { Copy } from "lucide-react";
 
 type CopyButtonProps = {
