+++ src/constants/urls.ts
@@ -1,0 +1,1 @@
+export const GITHUB_URL = "https://github.com/SuicaLondon/suica-no-toy-box";
