--- src/lib/components/providers/index.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { ReactNode, useState } from "react";
-
-export function Providers({ children }: { children: ReactNode }) {
-    const [queryClient] = useState(() => new QueryClient());
-
-    return (
-        <QueryClientProvider client={queryClient}>
-            {children}
-        </QueryClientProvider>
-    );
-}
