--- src/lib/constants/urls.ts
@@ -1,1 +1,0 @@
-export const GITHUB_URL = "https://github.com/SuicaLondon/suica-no-toy-box";
